// applets/radarr/api/model/movieResource.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Radarr
 * Radarr API docs
 * OpenAPI spec version: 3.0.0
 */
import type { Language } from "./language";
import type { AlternativeTitleResource } from "./alternativeTitleResource";
import type { MovieStatusType } from "./movieStatusType";
import type { MediaCover } from "./mediaCover";
import type { AddMovieOptions } from "./addMovieOptions";
import type { Ratings } from "./ratings";
import type { MovieFileResource } from "./movieFileResource";
import type { MovieCollection } from "./movieCollection";

export interface MovieResource {
  id?: number;
  title?: string | null;
  originalTitle?: string | null;
  originalLanguage?: Language;
  alternateTitles?: AlternativeTitleResource[] | null;
  secondaryYear?: number | null;
  secondaryYearSourceId?: number;
  sortTitle?: string | null;
  sizeOnDisk?: number | null;
  status?: MovieStatusType;
  overview?: string | null;
  inCinemas?: string | null;
  physicalRelease?: string | null;
  digitalRelease?: string | null;
  physicalReleaseNote?: string | null;
  images?: MediaCover[] | null;
  website?: string | null;
  remotePoster?: string | null;
  year?: number;
  hasFile?: boolean;
  youTubeTrailerId?: string | null;
  studio?: string | null;
  path?: string | null;
  qualityProfileId?: number;
  monitored?: boolean;
  minimumAvailability?: MovieStatusType;
  isAvailable?: boolean;
  folderName?: string | null;
  runtime?: number;
  cleanTitle?: string | null;
  imdbId?: string | null;
  tmdbId?: number;
  titleSlug?: string | null;
  rootFolderPath?: string | null;
  folder?: string | null;
  certification?: string | null;
  genres?: string[] | null;
  tags?: number[] | null;
  added?: string;
  addOptions?: AddMovieOptions;
  ratings?: Ratings;
  movieFile?: MovieFileResource;
  collection?: MovieCollection;
  popularity?: number;
}